requirements.txt
requests==2.32.3
Django>=4.0
patool>=4.0.0